requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
email-validator
python-dotenv
Flask-OAuthlib
requests-oauthlib
pymysql
Werkzeug